requirements.txt
pyodbc>=4.0.30
psycopg2-binary>=2.9.0
pyyaml>=5.4
